+++ offer-manager-app/src/main/groovy/life/qbic/portal/offermanager/components/offer/create/OfferOverviewView.groovy
@@ -3,23 +3,14 @@ package life.qbic.portal.offermanager.components.offer.create
 import com.vaadin.data.provider.ListDataProvider
 import com.vaadin.icons.VaadinIcons
 import com.vaadin.shared.ui.grid.HeightMode
-import com.vaadin.ui.Alignment
-import com.vaadin.ui.Button
-import com.vaadin.ui.Component
-import com.vaadin.ui.Grid
-import com.vaadin.ui.HorizontalLayout
-import com.vaadin.ui.Label
-import com.vaadin.ui.Panel
-import com.vaadin.ui.TextArea
-import com.vaadin.ui.VerticalLayout
+import com.vaadin.ui.*
 import com.vaadin.ui.components.grid.HeaderRow
 import com.vaadin.ui.renderers.NumberRenderer
 import com.vaadin.ui.themes.ValoTheme
 import groovy.util.logging.Log4j2
-import life.qbic.datamodel.dtos.business.services.Product
 import life.qbic.business.offers.Currency
+import life.qbic.datamodel.dtos.business.services.Product
 import life.qbic.portal.offermanager.components.GridUtils
-import life.qbic.portal.offermanager.dataresources.offers.OfferResourcesService
 
 /**
  * This class generates a Layout in which the user
@@ -29,11 +20,10 @@ import life.qbic.portal.offermanager.dataresources.offers.OfferResourcesService
  * with the intention of allowing the user to get an overview over all information he has entered before he creates and stores an
  * offer in the QBiC database.
  *
- * @since: 0.1.0
- *
+ * @since 0.1.0
  */
 @Log4j2
-class OfferOverviewView extends VerticalLayout{
+class OfferOverviewView extends VerticalLayout {
 
     private final CreateOfferViewModel createOfferViewModel
 
@@ -43,7 +33,7 @@ class OfferOverviewView extends VerticalLayout{
     Button save
     CreateOfferController createOfferController
 
-    OfferOverviewView(CreateOfferViewModel viewModel, CreateOfferController controller, OfferResourcesService service){
+    OfferOverviewView(CreateOfferViewModel viewModel, CreateOfferController controller) {
         this.createOfferViewModel = viewModel
         this.createOfferController = controller
         initLayout()
@@ -72,7 +62,7 @@ class OfferOverviewView extends VerticalLayout{
     /**
      * Initializes the start layout for this view
      */
-    private void initLayout(){
+    private void initLayout() {
         this.previous = new Button(VaadinIcons.CHEVRON_CIRCLE_LEFT)
         previous.addStyleName(ValoTheme.LABEL_LARGE)
 
@@ -90,7 +80,7 @@ class OfferOverviewView extends VerticalLayout{
 
         this.itemGrid = new Grid<>("Selected Items:")
 
-        this.addComponents(offerOverview,buttonLayout)
+        this.addComponents(offerOverview, buttonLayout)
         this.setMargin(false)
     }
 
@@ -113,7 +103,7 @@ class OfferOverviewView extends VerticalLayout{
 
             //specify size of grid and layout
             grid.setWidthFull()
-            grid.setHeightMode(HeightMode.UNDEFINED)
+            grid.setHeightByRows(6)
 
         } catch (Exception e) {
             new Exception("Unexpected exception in building the product item grid", e)
@@ -123,7 +113,7 @@ class OfferOverviewView extends VerticalLayout{
     /**
      * Fills the information of the offer into the panel
      */
-    void fillPanel(){
+    void fillPanel() {
         /*
         Container for the complete overview
          */
@@ -163,12 +153,10 @@ class OfferOverviewView extends VerticalLayout{
         TextArea experimentInfo = new TextArea("Experimental design")
         experimentInfo.setIcon(VaadinIcons.NOTEBOOK)
         experimentInfo.setValue(
-                createOfferViewModel.experimentalDesign ? createOfferViewModel.experimentalDesign: "No design defined.")
+                createOfferViewModel.experimentalDesign ? createOfferViewModel.experimentalDesign : "No design defined.")
         experimentInfo.setEnabled(false)
         experimentInfo.setWidth("100%")
         projectInfoForm.addComponent(experimentInfo)
-
-
 
 
         /*
@@ -206,8 +194,9 @@ class OfferOverviewView extends VerticalLayout{
                 new PriceField("Total Price", createOfferViewModel.totalPrice)
         ])
         gridLayout.addColumn(PriceField::getName)
-        gridLayout.addColumn(  {
-            costs -> costs.value},
+        gridLayout.addColumn({
+            costs -> costs.value
+        },
                 new NumberRenderer(Currency.getFormatterWithSymbol()))
 
         gridLayout.headerVisible = false
@@ -220,6 +209,7 @@ class OfferOverviewView extends VerticalLayout{
     Small helper object, that will display information
     about individual price positions for offer overviews.
      */
+
     private class PriceField {
 
         String name